PEP8 Changes
