+++ setup.py
@@ -1,4 +1,5 @@
 import ndio
+from distutils.core import setup
 
 VERSION = ndio.version
 """
@@ -14,23 +15,23 @@ git push --tags
 python setup.py sdist upload -r pypi
 """
 
-from distutils.core import setup
 setup(
-    name = 'ndio',
-    packages = [
+    name='ndio',
+    packages=[
         'ndio',
         'ndio.convert',
         'ndio.ramon',
         'ndio.remote',
         'ndio.utils'
     ],
-    version = VERSION,
-    description = 'A Python library for open neuroscience data access and manipulation.',
-    author = 'Jordan Matelsky',
-    author_email = 'jordan@neurodata.io',
-    url = 'https://github.com/neurodata/ndio',
-    download_url = 'https://github.com/neurodata/ndio/tarball/' + VERSION,
-    keywords = [
+    version=VERSION,
+    description='A Python library for open neuroscience data access and \
+manipulation.',
+    author='Jordan Matelsky',
+    author_email='jordan@neurodata.io',
+    url='https://github.com/neurodata/ndio',
+    download_url='https://github.com/neurodata/ndio/tarball/' + VERSION,
+    keywords=[
         'brain',
         'medicine',
         'microscopy',
@@ -50,11 +51,11 @@ setup(
         'fMRI',
         'calcium'
     ],
-    classifiers = [],
-    setup_requires = [
+    classifiers=[],
+    setup_requires=[
         'numpy',
     ],
-    install_requires = [
+    install_requires=[
         "pillow",
         "numpy",
         "h5py",
